// src/app/models/Quality.model.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface Quality {
    FarmCode: string;
    ReceptionDate: string;
    ChestContusions: string;
    Wings: string;
    Legs: string;
    BackScratches: string;
    MachineDamage: string;
    ManufactureExecution: string;
    Diagnostic: string;
    Reasons: string;
    Procedure: string;

}